consolidates data abstractions in model.py
