--- node.py
@@ -1,5 +1,0 @@
-class Node:
-
-    def __init__(self, state, ID):
-        self.state = state
-        self.ID = ID
