--- player.py
@@ -1,7 +1,0 @@
-class Player:
-
-    def __init__(self, mode, piece):
-
-        self.mode = mode
-        self.piece = piece
-        self.opponent = None
